Chronicle #6: Router + weighted_rrf_normalized fusion (final epic ticket)
Closes the Chronicle epic. SIMPLE-classified queries skip the BM25 + entity
channels (~50ms each) while always preserving the temporal channel (chronicle's
differentiator). Fusion swaps rank-only RRF for weighted RRF with per-channel
min-max score normalisation so the BM25-vs-cosine score-scale mismatch no
longer skews fused ordering.

Router relocated from khora.engines.vectorcypher.router to khora.query.router
(shared primitive); the vectorcypher path remains as a back-compat shim.
